Add ProxyProcessor Eclipse class loading test
